BAU: Use FindOneAndUpdate when inserting new PollingItems (#19)
There is a risk that a race condition could occur between finding and inserting a new polling item into the database.

This changes the logic to leverage Mongo's SetOnInsert which allows setting of the fields only on creation.  If the item does not exist in the database the function returns null by default, otherwise it returns the document.